+++ lib/app/core/widgets/widgets.dart
@@ -40,3 +40,12 @@ Widget buildDashboardWidget(final Widget child) {
     child: child,
   );
 }
+
+const double widthThresholdSmall = 600;
+const double widthThresholdMedium = 1200;
+
+extension ViewExtension on BuildContext {
+  bool get isWidthSmall => (MediaQuery.of(this).size.width <= widthThresholdSmall);
+  bool get isWidthMedium => (MediaQuery.of(this).size.width <= widthThresholdMedium);
+  bool get isWidthLarge => (MediaQuery.of(this).size.width > widthThresholdMedium);
+}