+++ lib/app/modules/home/sub_views/view_aliases/view_aliases.dart
@@ -2,6 +2,7 @@ import 'package:money/app/controller/selection_controller.dart';
 import 'package:money/app/core/helpers/list_helper.dart';
 import 'package:money/app/data/models/money_objects/aliases/alias.dart';
 import 'package:money/app/data/models/money_objects/transactions/transaction.dart';
+import 'package:money/app/data/storage/data/data.dart';
 import 'package:money/app/modules/home/sub_views/adaptive_view/adaptive_list/transactions/list_view_transactions.dart';
 import 'package:money/app/modules/home/sub_views/adaptive_view/view_money_objects.dart';