+++ lib/app/modules/home/sub_views/view_rentals/view_rentals.dart
@@ -4,6 +4,7 @@ import 'package:money/app/data/models/money_objects/rent_buildings/rent_building
 import 'package:money/app/data/models/money_objects/rental_unit/rental_unit.dart';
 import 'package:money/app/data/models/money_objects/splits/money_split.dart';
 import 'package:money/app/data/models/money_objects/transactions/transaction.dart';
+import 'package:money/app/data/storage/data/data.dart';
 import 'package:money/app/modules/home/sub_views/adaptive_view/adaptive_list/transactions/list_view_transactions.dart';
 import 'package:money/app/modules/home/sub_views/adaptive_view/view_money_objects.dart';
 import 'package:money/app/modules/home/sub_views/view_rentals/rental_pnl.dart';